requirements.txt
pinecone-client
openai>0.27.0
gunicorn>=20.0,<21.0
requests
flask
streamlit
